Menu model has been converted to size_t.
Chromium change:

https://source.chromium.org/chromium/chromium/src/+/283ef6e7432537f3c5fca19b642603ef62e762bc

commit 283ef6e7432537f3c5fca19b642603ef62e762bc
Author: Peter Kasting <pkasting@chromium.org>
Date:   Wed Jul 20 18:17:47 2022 +0000

    Reland "Convert MenuModel to size_t."

    This reverts commit aea7d02dd273a7cb391bfac65b0e1ed4eeaec08c.

    Bug: 1292951

+++ browser/ui/views/sidebar/sidebar_control_view.cc
@@ -39,7 +39,7 @@ class ControlViewMenuModel : public ui::SimpleMenuModel {
   ControlViewMenuModel& operator=(const ControlViewMenuModel&) = delete;
 
   // ui::SimpleMenuModel overrides:
-  const gfx::FontList* GetLabelFontListAt(int index) const override {
+  const gfx::FontList* GetLabelFontListAt(size_t index) const override {
     if (index == 0) {
       return &ui::ResourceBundle::GetSharedInstance().GetFontList(
           ui::ResourceBundle::BoldFont);
